initial commit for e2e test
The test does not yet do any verification of tlp-stress running. So far it
checks that a casskop CassandraCluster is created and initialized.

There are a bunch of setup/teardown steps that need to be automated. For
example, casskop needs to be installed. This includes CRDs and then the
operator itself.